+++ mlbpestimation/vitaldb/casesplit.py
@@ -2,8 +2,7 @@ from typing import List
 from zlib import crc32
 
 import numpy as np
-from numpy import float64
-from tensorflow import TensorSpec
+from tensorflow import TensorSpec, float32
 from tensorflow.python.data import Dataset
 
 from mlbpestimation.data.multipartdataset import MultipartDataset
@@ -26,7 +25,7 @@ def load_vitaldb_dataset():
             Dataset.from_generator(
                 lambda c=case_split: VitalDBGenerator(options, DatasetApi(), c),
                 output_signature=(
-                    TensorSpec(shape=(None, 1), dtype=float64)
+                    TensorSpec(shape=(None, 1), dtype=float32)
                 )
             )
         )
